Move let binding to where clause
Seems like a good idea to have both of the simple bindings be in the
where clause, instead of having them split. An argument could also be
made for both to be let statements, but I prefer this way.

Also aligned the where clause to the proper fuction level. It was a bit
too far over, even if it was still working.